+++ CHANGELOG.md
@@ -1,14 +1,21 @@
 ## CHANGELOG
 
+### 3.4.0 (2021-06-XX)
+**SDK**
+- **Fix:** Prevent exception during calling `onError` asynchronously in `getUniqueId`.
+
 ### 3.3.0 (2021-05-19)
 **SDK**
 - **Deprecated:** Old `getUniqueId`interface.
 - **Feature:** Added `getUniqueId` new interface for invoking data using `onSuccess` and `onError`.
 - **Feature:** Support `mailto` uri address. See [this](https://developer.mozilla.org/en-US/docs/Web/HTML/Element/a).
 - **Feature:** Added `rakuten.miniapp.user.action.SEND_MESSAGE` custom permission and applied to `ChatBridgeDispatcher.sendMessageToContactId`.
+- **Change:** Added `getAccessToken` result with new Error type i.e `MiniAppAccessTokenError` to support predefined error types, previous `getAccessToken` is deprecated and can not be used anymore.
+- **Feature:** Update `MiniAppSdkConfig` class to accept list of `MiniAppAnalyticsConfig` to send analytics in multiple account.
 
 **Sample App**
 - **Change:** Replaced the implementation of `getUniqueId` using new interface.
+- **Feature:** Added QA option in settings screen to test `authorizationFailureError` and `custom` error type.
 
 ### 3.2.0 (2021-04-23)
 **SDK**